Refactored ClientModule repositories
