+++ lib/modules/clients/data/repository/sqlite_client_repository.dart
@@ -1,16 +1,12 @@
 import 'package:fpdart/fpdart.dart';
 import 'package:sqflite/sqflite.dart';
 
+import '../../clients.dart';
 import 'sqlite_address_repository.dart';
 import 'sqlite_contact_repository.dart';
 import '../../../../core/failure.dart';
 import '../../../../extensions.dart';
 import '../../../../database/sqlite/sqlite.dart';
-import '../../domain/dto/listing_client_dto.dart';
-import '../../domain/model/client.dart';
-import '../../domain/model/address.dart';
-import '../../domain/model/contact.dart';
-import '../../domain/repository/client_repository.dart';
 
 class SQLiteClientRepository extends SQLiteRepository<Client>
     implements ClientRepository {
