+++ lib/modules/clients/data/repository/sqlite_contact_repository.dart
@@ -1,16 +1,12 @@
-import 'package:equatable/equatable.dart';
 import 'package:fpdart/fpdart.dart';
-import 'package:json_annotation/json_annotation.dart';
 import 'package:sqflite/sqflite.dart';
 
+import '../../clients.dart';
 import '../../../../core/core.dart';
-import '../../../../database/database.dart';
 import '../../../../database/sqlite/sqlite.dart';
-import '../../domain/model/contact.dart';
 
-part 'sqlite_contact_repository.g.dart';
-
-class SQLiteContactRepository extends SQLiteRepository<ContactEntity> {
+class SQLiteContactRepository extends SQLiteRepository<ContactEntity>
+    implements ContactRepository {
   SQLiteContactRepository(SQLiteDatabase database)
       : super(
           'clientContacts',
@@ -42,28 +38,4 @@ class SQLiteContactRepository extends SQLiteRepository<ContactEntity> {
       return Left(DatabaseFailure(SQLiteRepository.couldNotDeleteMessage, e));
     }
   }
-}
-
-@JsonSerializable()
-class ContactEntity extends Equatable implements Entity<int> {
-  @override
-  final int? id;
-  final int? clientId;
-  final String contact;
-
-  const ContactEntity({this.id, required this.contact, this.clientId});
-
-  ContactEntity.fromContact(Contact contact, {this.clientId})
-      : id = null,
-        contact = contact.contact;
-
-  factory ContactEntity.fromJson(Map<String, dynamic> json) =>
-      _$ContactEntityFromJson(json);
-
-  Map<String, dynamic> toJson() => _$ContactEntityToJson(this);
-
-  Contact toContact() => Contact(contact: contact);
-
-  @override
-  List<Object?> get props => [id, clientId, contact];
 }
