+++ lib/modules/clients/domain/domain.dart
@@ -3,5 +3,7 @@ export 'model/client.dart';
 export 'model/contact.dart';
 export 'model/states.dart';
 export 'repository/client_repository.dart';
+export 'repository/address_repository.dart';
+export 'repository/contact_repository.dart';
 export 'usecase/usecase.dart';
 export 'dto/listing_client_dto.dart';
