+++ test/modules/clients/data/repository/sqlite_contact_repository_test.dart
@@ -1,4 +1,5 @@
 import 'package:kitchen_helper/database/sqlite/sqlite.dart';
+import 'package:kitchen_helper/modules/clients/clients.dart';
 import 'package:kitchen_helper/modules/clients/data/repository/sqlite_contact_repository.dart';
 import 'package:mocktail/mocktail.dart';
 import 'package:test/test.dart';
